add filter by budget and fix some error with handling big numbers to the edtText

+++ app/src/main/java/com/abdullah996/leadscrm/network/Apis.kt
@@ -164,8 +164,8 @@ interface Apis {
     // filter by budget
     @GET("lead/filter")
     suspend fun filterByBudget(
-        @Query("budget_from") from: Int,
-        @Query("budget_to") to: Int
+        @Query("budget_from") from: Long,
+        @Query("budget_to") to: Long
     ):Response<LeedsReponse>
 
     // filter by unit Type id
